requirements.txt
alembic==1.8.0
anyio==3.6.1
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.15.0
websockets==10.3
zipp==3.8.1
